Restructured
Learned a bit more about lua and how to separate into multiple files.
We have a working version of the plugin.

Currently, I have the whole directory simlinked to the Reaper Scripts
directory for testing.  I also have a wrapper over the Scythe GUI
library to add different screens to the plugin a bit easier.